feat(cb2-5213): reformat unit tests
