+++ tests/unit/send.test.ts
@@ -6,6 +6,7 @@ import { PutEventsResponse, PutEventsRequest, PutEventsResultEntry } from 'aws-s
 import { SendResponse } from '../../src/eventbridge/SendResponse';
 import { TestActivity } from '../../src/utils/testActivity';
 import { sendEvents } from '../../src/eventbridge/send';
+import { Differences } from '../../src/utils/differences';
 
 jest.mock('aws-sdk', () => {
   const mEventBridgeInstance = {
@@ -28,13 +29,13 @@ const mPutEventsResponse = { Entries: Array<PutEventsResultEntry>(1) };
 // eslint-disable-next-line @typescript-eslint/unbound-method
 mocked(mEventBridgeInstance.putEvents as PutEventsWithParams).mockImplementation(
   (params: PutEventsRequest): AWS.Request<PutEventsResponse, AWS.AWSError> => {
-    if (params.Entries[0].Detail.includes('\\"testResultId\\":\\"HandledError\\"')) {
+    if (params.Entries[0].Detail.includes('\\":\\"HandledError\\"')) {
       mPutEventsResponse.Entries[0] = { ErrorMessage: 'Failed to process event.' };
     } else {
       mPutEventsResponse.Entries[0] = { EventId: '2468' };
     }
 
-    if (params.Entries[0].Detail.includes('\\"testResultId\\":\\"UnhandledError\\"')) {
+    if (params.Entries[0].Detail.includes('\\":\\"UnhandledError\\"')) {
       mResultInstance.promise = jest.fn().mockReturnValue(Promise.reject(new Error('Oh no!')));
     } else {
       mResultInstance.promise = jest.fn().mockReturnValue(Promise.resolve(mPutEventsResponse));
@@ -45,20 +46,20 @@ mocked(mEventBridgeInstance.putEvents as PutEventsWithParams).mockImplementation
 );
 
 describe('Send events', () => {
-  describe('Events sent', () => {
+  describe('Events sent as test activities', () => {
     // @ts-ignore
     const consoleSpy = jest.spyOn(console._stdout, 'write');
 
-    it('GIVEN one event to send WHEN sent THEN one event is returned.', async () => {
+    it('GIVEN one activities event to send WHEN sent THEN one event is returned.', async () => {
       const mTestResult: TestActivity[] = [createTestResult()];
       const mSendResponse: SendResponse = { SuccessCount: 1, FailCount: 0 };
       await expect(sendEvents(mTestResult, 'completion')).resolves.toEqual(mSendResponse);
     });
 
-    it('GIVEN two events to send WHEN sent THEN two events are returned and two infos logged.', async () => {
+    it('GIVEN two activities events to send WHEN sent THEN two events are returned and two infos logged.', async () => {
       const mTestResult: TestActivity[] = [createTestResult('Result1'), createTestResult('Result2')];
       const mSendResponse: SendResponse = { SuccessCount: 2, FailCount: 0 };
-      await expect(sendEvents(mTestResult)).resolves.toEqual(mSendResponse);
+      await expect(sendEvents(mTestResult, 'completion')).resolves.toEqual(mSendResponse);
       expect(consoleSpy).toHaveBeenCalledWith(
         expect.stringContaining(
           `info: Result sent to eventbridge (testResultId: 'Result1', vin: '${mTestResult[0].vin}')`,
@@ -71,18 +72,18 @@ describe('Send events', () => {
       );
     });
 
-    it('GIVEN an event WHEN eventbridge could not process it THEN log info message and error', async () => {
+    it('GIVEN an activities event WHEN eventbridge could not process it THEN log info message and error', async () => {
       const mTestResult: TestActivity[] = [createTestResult('HandledError')];
-      await sendEvents(mTestResult);
+      await sendEvents(mTestResult, 'completion');
       expect(consoleSpy).toHaveBeenCalledWith(
         expect.stringContaining(
-          `info: Failed to send result to eventbridge (testResultId: '${mTestResult[0].testResultId}', vin: '${mTestResult[0].vin}')`,
+          `info: Failed to send to eventbridge (testResultId: '${mTestResult[0].testResultId}', vin: '${mTestResult[0].vin}')`,
         ),
       );
       expect(consoleSpy).toHaveBeenCalledWith(expect.stringContaining('Error: Failed to process event.'));
     });
 
-    it('GIVEN an issue with eventbridge WHEN 5 events are sent and 1 fails THEN the failure is in the response and an error is logged.', async () => {
+    it('GIVEN an issue with eventbridge WHEN 5 activities events are sent and 1 fails THEN the failure is in the response and an error is logged.', async () => {
       const mTestResult: TestActivity[] = [
         createTestResult(),
         createTestResult(),
@@ -91,8 +92,34 @@ describe('Send events', () => {
         createTestResult('UnhandledError'),
       ];
       const mSendResponse: SendResponse = { SuccessCount: 4, FailCount: 1 };
-      await expect(sendEvents(mTestResult)).resolves.toEqual(mSendResponse);
+      await expect(sendEvents(mTestResult, 'completion')).resolves.toEqual(mSendResponse);
       expect(consoleSpy).toHaveBeenCalledWith(expect.stringContaining('Error: Oh no!'));
+    });
+  });
+
+  describe('Events sent as differences', () => {
+    it('GIVEN one differences event to send WHEN sent THEN one event is returned.', async () => {
+      const mDifferences: Differences[] = [createDifferences(1)];
+      const mSendResponse: SendResponse = { SuccessCount: 1, FailCount: 0 };
+      await expect(sendEvents(mDifferences, 'amendment')).resolves.toEqual(mSendResponse);
+    });
+
+    it('GIVEN two differences events to send WHEN sent THEN two events are returned.', async () => {
+      const mDifferences: Differences[] = [createDifferences(1), createDifferences(2)];
+      const mSendResponse: SendResponse = { SuccessCount: 2, FailCount: 0 };
+      await expect(sendEvents(mDifferences, 'amendment')).resolves.toEqual(mSendResponse);
+    });
+
+    it('GIVEN an issue with eventbridge WHEN 6 differences events are sent and 1 fails THEN the failure is in the response.', async () => {
+      const mDifferences: Differences[] = [
+        createDifferences(1),
+        createDifferences(1),
+        createDifferences(1),
+        createDifferences(1),
+        createDifferences(1, 'HandledError'),
+      ];
+      const mSendResponse: SendResponse = { SuccessCount: 4, FailCount: 1 };
+      await expect(sendEvents(mDifferences, 'amendment')).resolves.toEqual(mSendResponse);
     });
   });
 });
@@ -116,4 +143,16 @@ function createTestResult(resultId?: string): TestActivity {
     testResultId: resultId || '1234',
   };
   return activityEvent;
+}
+
+function createDifferences(qty: number, reasonForCreation = 'foo'): Differences {
+  const fields = [];
+  for (let i = 0; i < qty; i++) {
+    fields.push({
+      fieldname: Math.random().toString(),
+      oldvalue: Math.random().toString(),
+      newValue: Math.random().toString(),
+    });
+  }
+  return { reason: reasonForCreation, fields };
 }
